upgrade the types and add functions to edit the quantity in cart page
