+++ package.json
@@ -15,6 +15,7 @@
     "cors": "^2.8.5",
     "cross-env": "^7.0.3",
     "express": "^4.18.2",
+    "helmet": "^7.1.0",
     "jspdf": "^2.5.1",
     "pdfkit": "^0.15.0",
     "react": "^18.2.0",
