+++ src/client/pages/CollectionDetailPage.tsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
+import { CollectionProps } from "../types/types";
 import axios from "axios";
 
 import './CollectionDetailPage.css';
@@ -9,7 +10,7 @@ const PAGE_SIZE = 3;
 const CollectionDetailPage: React.FC = () => {
     const history = useNavigate();
     const { key } = useParams();
-    const [data, setData] = useState<any>(null);
+    const [data, setData] = useState<CollectionProps | null>(null);
     const [page, setPage] = useState(1);
     const [total, setTotal] = useState(1);
 
@@ -39,10 +40,10 @@ const CollectionDetailPage: React.FC = () => {
                     <div className="flex flex-row w-full items-center justify-between px-5">
                         {data.products
                         ?.filter(
-                        (_: any, index: number) =>
+                        (_, index) =>
                             index >= (page - 1) * PAGE_SIZE && index < page * PAGE_SIZE
                         )
-                        .map((product: any) => {
+                        .map((product) => {
                         return (
                             <div className="flex flex-col items-center justify-content">
                                 <img
