+++ src/server/main.ts
@@ -2,247 +2,13 @@ import express from "express";
 import bodyParser from "body-parser";
 import cors from 'cors';
 import ViteExpress from "vite-express";
-import axios from "axios";
+import router from "./routes";
 
 const app = express();
+
 app.use(cors());
 app.use(bodyParser.json());
-
-let collections = [
-  {
-      "title": "Baklava",
-      "description": "Baklava is a sweet dessert that is popular in Middle Eastern and Mediterranean cuisine. Our baklava is made from layers of thin and delicate filo pastry that are filled with chopped nuts, such as pistachios, walnuts, or almonds, and held together with syrup.",
-      "image": "https://www.kermanig.com/cdn/shop/files/21.png?v=1681339473&width=750",
-      "products": [16,17,18,19]
-  },
-  {
-      "title": "Focaccia",
-      "description": "Every Focaccia leaving our bakery is made by hand using only the finest ingredients.",
-      "image": "https://www.kermanig.com/cdn/shop/files/template-focaccias_trio.jpg?v=1689358976&width=750",
-      "products": [0,1,2,3]
-  },
-  {
-      "title": "Pita Chips",
-      "description": "Our Pita chips are a great crunchy snack item that are made from small pieces of pita bread that have been perfectly seasoned.",
-      "image": "https://www.kermanig.com/cdn/shop/files/new_pita_trio_1.png?v=1689358983&width=750",
-      "products": [4,5,6,7]
-  },
-  {
-      "title": "Cakes",
-      "description": "Satisfy your sweet tooth with our delicious cakes and cupcake",
-      "image": "https://www.kermanig.com/cdn/shop/files/CUPCAKE_TRIO_23232.png?v=1689359235&width=750",
-      "products": [8,9,10,11]
-  },
-  {
-      "title": "Cookies and Pastries",
-      "description": "Freshly baked cookies and pastries with a mouth-watering taste.",
-      "image": "https://www.kermanig.com/cdn/shop/files/Cookie_Trio_Pic.png?v=1689363652&width=750",
-      "products": [12,13,14,15]
-  }
-];
-
-let products = [
-  {
-      "title": "Mozzarella & Pesto Focaccia",
-      "image": "https://www.kermanig.com/cdn/shop/products/Pesto1.jpg?v=1595713723&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": "Mozzarella & Basil Focaccia",
-      "image": "https://www.kermanig.com/cdn/shop/products/Mozzarella1.jpg?v=1595713620&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": " Rustic Olive Focaccia",
-      "image": "https://www.kermanig.com/cdn/shop/products/20200615_170657.jpg?v=1595713632&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": "Jalapeño & Cheese Focaccia",
-      "image": "https://www.kermanig.com/cdn/shop/products/20200615_172508.jpg?v=1596138274&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": "Pita Chips - Garlic",
-      "image": "https://www.kermanig.com/cdn/shop/products/Pitachipsalone_fbd05619-89f1-47fa-8e0f-b391016d20b0.jpg?v=1595536152&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": "Pita Chips - Lemon & Herbs",
-      "image": "https://www.kermanig.com/cdn/shop/products/Pitachipsalone.jpg?v=1595535980&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": "Pita Chips - Spicy",
-      "image": "https://www.kermanig.com/cdn/shop/products/Pitachipsalone_3ff33094-f272-49a1-9cc5-cc51aabf6755.jpg?v=1595536069&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": "Pita chips - Original",
-      "image": "https://www.kermanig.com/cdn/shop/products/Pitachipsalone_48d0702b-a7fb-4cfb-9861-2776dce020ba.jpg?v=1680734618&width=360",
-      "price": 7.99,
-      "star": 3
-  },
-  {
-      "title": "Almond Cake",
-      "image": "https://www.kermanig.com/cdn/shop/products/20200701_161708.jpg?v=1596140081&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Chocolate Chip Cake",
-      "image": "https://www.kermanig.com/cdn/shop/products/20200701_161144.jpg?v=1600892685&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Raisin Cake",
-      "image": "https://www.kermanig.com/cdn/shop/products/20200701_161559.jpg?v=1596141020&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Cup Cakes",
-      "image": "https://www.kermanig.com/cdn/shop/products/20200701_160923.jpg?v=1596141214&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Coconut Pastry",
-      "image": "https://www.kermanig.com/cdn/shop/files/Coconut.jpg?v=1696628551&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "French Cookies",
-      "image": "https://www.kermanig.com/cdn/shop/products/2.jpg?v=1695841527&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Lemon Cookie",
-      "image": "https://www.kermanig.com/cdn/shop/products/LemonCookieOnPlate.png?v=1695840875&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Chocolate Chip Cookie",
-      "image": "https://www.kermanig.com/cdn/shop/products/Chocolatechipcookieedited.png?v=1612909612&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Long Baklava",
-      "image": "https://www.kermanig.com/cdn/shop/products/2222.jpg?v=1681316022&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Square Baklava",
-      "image": "https://www.kermanig.com/cdn/shop/products/sqauresngle111.jpg?v=1668119124&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Round Baklava",
-      "image": "https://www.kermanig.com/cdn/shop/products/RoundSinglepicture202311111.jpg?v=1668106734&width=360",
-      "price": 9.99,
-      "star": 3
-  },
-  {
-      "title": "Baklava Assorted",
-      "image": "https://www.kermanig.com/cdn/shop/products/Assorted_3.jpg?v=1694804397&width=360",
-      "price": 9.99,
-      "star": 3
-  }
-];
-
-let carousels = [
-  {
-      "src": "//www.kermanig.com/cdn/shop/files/Sarkis.jpg?v=1681245557&width=3840",
-      "height": "1352",
-      "loading": "lazy",
-      "sizes": "120vw"
-  },
-  {
-      "src": "//www.kermanig.com/cdn/shop/files/Orange_and_Green_Abstract_Shapes_Free_Shipping_Nationwide_Banner_09ab6816-8a57-4217-9d47-990a6742ed68.png?v=1684540685&amp;width=3840",
-      "height": "2500",
-      "loading": "lazy",
-      "sizes": "120vw"
-  },
-  {
-      "src": "//www.kermanig.com/cdn/shop/files/Monthly_promotions_4.jpg?v=1684538884&amp;width=3840",
-      "alt": "",
-      "height": "1352",
-      "loading": "lazy",
-      "sizes": "120vw"
-  }
-];
-app.get('/api/:type', async (req, res, next) => {
-  try {
-    let data;
-    switch (req.params.type) {
-      case 'collections': 
-        data = collections; 
-        break;
-      case 'products': 
-        data = products; 
-        break;
-      case 'carousels': 
-        data = carousels; 
-        break;
-      default: break;
-    }
-    res.status(200).json(data)
-  }
-  catch (error) {
-    console.error('Error while parsing the data', error);
-  }
-})
-
-app.post('/api/:type', async (req, res, next) => {
-  try {
-    const { data } = req.body;
-    switch (req.params.type) {
-      case 'collections': 
-        collections = JSON.parse(data); 
-        break;
-      case 'products': 
-        products = JSON.parse(data); 
-        break;
-      case 'carousels': 
-        carousels = JSON.parse(data); 
-        break;
-      default: break;
-    }
-    res.status(200).json({ "success": true })
-  }
-  catch (error) {
-    console.error('Error while parsing the data', error);
-  }
-})
-
-app.get('/api/collections/:title', async (req, res, next) => {
-  const collection = collections.find(a => a.title === req.params.title);
-  const fullInfo = {
-    ...collection,
-    products: collection?.products?.map(id => products.at(id))
-  };
-  res.status(200).json(fullInfo || null);
-})
-
-app.post("/api/image", async (req, res, next) => {
-  const { url } = req.body;
-  const response = await axios.get(url, { responseType: "arraybuffer"});
-  res.json(response.data);
-})
+app.use('/api', router);
 
 ViteExpress.listen(app, 3000, () =>
   console.log("Server is listening on port 3000...")
